+++ src/builtin/pick_os_string.rs
@@ -1,0 +1,11 @@
+use std::ffi::OsString;
+
+use crate::{PickerArgResult, SinglePickable};
+
+impl SinglePickable for OsString {
+    fn pick_single(str: Option<&str>) -> PickerArgResult<Self> {
+        str.map_or(PickerArgResult::NotFound, |s| {
+            PickerArgResult::Parsed(Self::from(s))
+        })
+    }
+}