+++ test/src/test.rs
@@ -6,5 +6,6 @@ mod pos_matcher_test;
 mod priority_test;
 mod route_test;
 mod style_test;
+mod value_extra_test;
 mod value_flag_test;
 mod value_string_test;